update sistence action
