+++ kairei-core/src/sistence/action/mod.rs
@@ -1,2 +1,3 @@
 pub mod error;
+pub mod model;
 pub mod service;
